+++ hardware/vga/vga_sequencer_textmode.c
@@ -102,7 +102,7 @@ void VGA_TextDecoder(VGA_Type *VGA, word loadedlocation)
 	else //VGA mode?
 	{
 	VGAtext: //VGA text catch-all!
-		attr3 = attribute; //Load the attribute!
+		attr3 = (byte)attribute; //Load the attribute!
 		attr3 >>= 3; //...
 		attr3 &= 1; //... Take bit 3 to get the actual attribute we need!
 		x = 0; //Start with the first pixel!
